+++ components/ChatMessageList.tsx
@@ -1,91 +1,13 @@
 import * as React from 'react';
 import { shallow } from 'zustand/shallow';
 
-import { AspectRatio, Box, Button, Grid, List, Stack, Textarea, Typography } from '@mui/joy';
+import { Box, List } from '@mui/joy';
 import { SxProps } from '@mui/joy/styles/types';
 
 import { ChatMessage } from '@/components/ChatMessage';
-import { DMessage, useActiveConfiguration, useActiveConversation, useChatStore } from '@/lib/store-chats';
-import { SystemPurposeId, SystemPurposes } from '@/lib/data';
+import { DMessage, useActiveConversation, useChatStore } from '@/lib/store-chats';
+import { PurposeSelector } from '@/components/PurposeSelector';
 import { useSettingsStore } from '@/lib/store';
-
-
-function PurposeSelect() {
-  const { setSystemPurposeId, systemPurposeId } = useActiveConfiguration();
-
-  const handlePurposeChange = (purpose: SystemPurposeId | null) => {
-    if (purpose) {
-      setSystemPurposeId(purpose);
-    }
-  };
-
-  function handleCustomSystemMessageChange(v: React.ChangeEvent<HTMLTextAreaElement>): void {
-    SystemPurposes['Custom'].systemMessage = v.target.value;
-  }
-
-  return (
-    <Stack direction='column' sx={{ justifyContent: 'center', alignItems: 'center', mx: 2, minHeight: '60vh' }}>
-
-      <Box>
-
-        <Typography level='body3' color='neutral' sx={{ mb: 2 }}>
-          AI purpose
-        </Typography>
-
-        <Grid container spacing={1}>
-          {Object.keys(SystemPurposes).map(spId => (
-            <Grid key={spId} xs={4} lg={3} xl={2}>
-              <AspectRatio variant='plain' ratio={1} sx={{
-                minWidth: { xs: 56, lg: 78, xl: 130 }, maxWidth: 130,
-                // borderRadius: 8,
-                // boxShadow: theme.vars.shadow.md,
-              }}>
-                <Button
-                  variant={systemPurposeId === spId ? 'solid' : 'soft'}
-                  color={systemPurposeId === spId ? 'primary' : 'neutral'}
-                  onClick={() => handlePurposeChange(spId as SystemPurposeId)}
-                  sx={{
-                    flexDirection: 'column',
-                    gap: { xs: 2, lg: 3 },
-                    /*fontFamily: theme.vars.fontFamily.code,*/ fontWeight: 500,
-                  }}
-                >
-                  <div style={{ fontSize: '1.875rem' }}>
-                    {SystemPurposes[spId as SystemPurposeId]?.symbol}
-                  </div>
-                  <div>
-                    {SystemPurposes[spId as SystemPurposeId]?.title}
-                  </div>
-                </Button>
-              </AspectRatio>
-            </Grid>
-          ))}
-        </Grid>
-
-        <Typography level='body2' sx={{ mt: 2 }}>
-          {SystemPurposes[systemPurposeId].description}
-        </Typography>
-
-        {systemPurposeId === 'Custom' && (
-          <>
-            <Textarea variant='soft' autoFocus placeholder={"Enter your custom system message here..."}
-            minRows={5} maxRows={12}
-            // onKeyDown={handleKeyPress}
-            // onDragEnter={handleMessageDragEnter}
-            defaultValue={SystemPurposes['Custom'].systemMessage} onChange={(e) => handleCustomSystemMessageChange(e)}
-            sx={{
-              fontSize: '16px',
-              lineHeight: 1.75,
-            }} />
-          </>
-        )}
-
-
-      </Box>
-
-    </Stack>
-  );
-}
 
 
 /**
@@ -114,7 +36,7 @@ export function ChatMessageList(props: { disableSend: boolean, sx?: SxProps, run
   // when there are no messages, show the purpose selector
   if (!filteredMessages.length) return (
     <Box sx={props.sx || {}}>
-      <PurposeSelect />
+      <PurposeSelector />
     </Box>
   );
